Created fuctioning game for one round at a time
Got a little carried away on a flight to Monterrey, Mexico  to play the
Monterrey Sultanes for a three game series and forgot to
update git. Made functions to get computer choice and human choice, and
declare a winner between the two. Also keeps track of score for the one
round. Playing five rounds and keeping track of score looks a little
tougher, which is what reminded me to update git.